+++ aa-display/src/main/java/io/github/nitsuya/aa/display/ui/aa/AaDisplayActivityKt.kt
@@ -14,10 +14,6 @@ object AaDisplayActivityKt {
         CoreApi.pressKey(action)
     }
 
-    fun toggleDisplayPower(){
-        CoreApi.toggleDisplayPower()
-    }
-
     fun toast(msg: String){
         CoreApi.toast(msg)
     }